[FEATURE] Compatibility with TYPO3 6.0+
The change adds a new namespaced class that contains basically
the same stuff than the previous class. All occurrences of t3lib_div
have been replaced with the new namespaced class.

In addition the Xclass-Inclusion in ext_localconf.php is modified
and the TYPO3 version dependency in ext_emconf.php is raised to
6.0+

So far it looks as if it's working - needs some deeper testing!