[FEATURE] Compatibility with TYPO3 6.0+
The change adds a new namespaced class that contains basically
the same stuff than the previous class. All occurrences of t3lib_div
have been replaced with the new namespaced class.

This change should not break compatibility with TYPO3 versions
below 6.0 as the old Xclass stuff is kept as is.

So far it looks as if it's working on 6.1-dev, but needs some
deeper testing!